pyproject.toml
[project]
name = "portfolio-fdc"
version = "0.1.0"
description = "Portfolio-friendly FDC toolkit (logger stream → segmentation → features → SPC dashboard)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.4.2",
  "pandas>=3.0.1",
  "PyYAML>=6.0",
  "requests>=2.32.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "ruff>=0.6",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "types-PyYAML>=6.0.12",
  "types-requests>=2.32.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "integration: end-to-end style tests combining generators and scraper",
  "slow: heavy tests that generate large files or run longer",
]

[tool.setuptools.packages.find]
where = ["src"]
